Swap price to float. try to count price in Weights.

+++ MiniShop/Models/Product.cs
@@ -7,7 +7,7 @@ namespace MiniShop.Models
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public decimal Price { get; set; }
+        public float Price { get; set; }
         public int Quantity { get; set; }
         public bool IsWeighable { get; set; }
         public int[] Weight { get; set; } = new int[2];
@@ -15,7 +15,7 @@ namespace MiniShop.Models
         public string PhotoPath { get; set; }
         public string DisplayInfo => $"{Name} — {Price}₽ × {Quantity}";
 
-        public Product(int id, string name, decimal price, int quantity, bool isWeighable,
+        public Product(int id, string name, float price, int quantity, bool isWeighable,
             int[] weight, DateTime expiryDate, string photoPath) 
         {
             Id = id;
